refactor: extract EDNS option codec into DNSpacket.EDNS
Mechanically move the EDNS option encoder (create_option_binary),
decoder (parse_opt_code) and their helpers (create_ecs_address_bytes,
pad_address, apply_prefix_mask) into a dedicated DNSpacket.EDNS
module, so all per-option wire-format knowledge lives in one place.

Public API is unchanged: create_edns_options/1 delegates to
EDNS.encode_options/1 and parse_opt_rr/2 keeps its signature.
